pudiste terminar la parte de los clientes y presupuesto, hay un error para hacer los de agregar productos

+++ Repositories/ProductoRepository.cs
@@ -1,6 +1,6 @@
 namespace EspacioRepositorios;
 
-using EspacioModelos;
+using EspacioModels;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.Sqlite;
@@ -60,7 +60,7 @@ public class ProductoRepository
                     Producto productoLeido = new Producto();
                     productoLeido.Id = Convert.ToInt32(lector["idProducto"]);
                     productoLeido.Descripcion = lector["Descripcion"].ToString();
-                    productoLeido.Precio = Convert.ToDouble(lector["Precio"]);
+                    productoLeido.Precio = Convert.ToDecimal(lector["Precio"]);
                     lista.Add(productoLeido);
                 }
             }
